Remove mute commands and logic from mod cog
