+++ redbot/cogs/mod/casetypes.py
@@ -56,48 +56,6 @@ CASETYPES = [
         "audit_type": "member_update",
     },
     {
-        "name": "vmute",
-        "default_setting": False,
-        "image": "\N{SPEAKER WITH CANCELLATION STROKE}",
-        "case_str": "Voice Mute",
-        "audit_type": "overwrite_update",
-    },
-    {
-        "name": "cmute",
-        "default_setting": False,
-        "image": "\N{SPEAKER WITH CANCELLATION STROKE}",
-        "case_str": "Channel Mute",
-        "audit_type": "overwrite_update",
-    },
-    {
-        "name": "smute",
-        "default_setting": True,
-        "image": "\N{SPEAKER WITH CANCELLATION STROKE}",
-        "case_str": "Server Mute",
-        "audit_type": "overwrite_update",
-    },
-    {
-        "name": "vunmute",
-        "default_setting": False,
-        "image": "\N{SPEAKER}",
-        "case_str": "Voice Unmute",
-        "audit_type": "overwrite_update",
-    },
-    {
-        "name": "cunmute",
-        "default_setting": False,
-        "image": "\N{SPEAKER}",
-        "case_str": "Channel Unmute",
-        "audit_type": "overwrite_update",
-    },
-    {
-        "name": "sunmute",
-        "default_setting": True,
-        "image": "\N{SPEAKER}",
-        "case_str": "Server Unmute",
-        "audit_type": "overwrite_update",
-    },
-    {
         "name": "vkick",
         "default_setting": False,
         "image": "\N{SPEAKER WITH CANCELLATION STROKE}",
